requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
django-crontab==0.7.1
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
iniconfig==1.1.1
Markdown==3.3.4
packaging==21.0
pluggy==1.0.0
psutil==5.8.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==3.0.3
pytest==6.2.5
pytz==2021.3
requests==2.26.0
sqlparse==0.2.4
toml==0.10.2
urllib3==1.26.7
